+++ packages/core-modules/test/contracts/modules/ElectionModule/Schedule.test.js
@@ -11,7 +11,7 @@ const {
 const { daysToSeconds } = require('@synthetixio/core-js/utils/misc/dates');
 const { bootstrap } = require('@synthetixio/deployer/utils/tests');
 const initializer = require('../../../helpers/initializer');
-const { ElectionPeriod, assertDatesAreClose } = require('../../../helpers/election-helper');
+const { ElectionPeriod, assertDatesAreClose } = require('./helpers/election-helper');
 const { findEvent } = require('@synthetixio/core-js/utils/ethers/events');
 
 describe('ElectionModule (schedule)', () => {
